requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic[email]==2.5.0
python-dotenv==1.0.0
PyJWT==2.7.0
bcrypt==4.1.1
password-validator==1.0
python-multipart==0.0.6
aiofiles==23.2.1
pydantic-settings==2.1.0
openpyxl==3.1.2
